Don't report missing sound files or wrong KHz background music.
It's fine for mappers (which should test their maps in /developer
anyways) but mostly just annoying when playing the game. Missing sounds
are mostly not available custom player models that spam the console on
connecting of other clients with non present sounds.

Also there are quite a few maps which have very short looping background
music tracks in not 22 KHz stereo what makes the console print a warning
on every single loop. That can result in a massive spam of useless
information since the engine handles the sounds very fine.