Improve word wrap.
Don't count the same word length after every advance, just do it once
and decide whether to split it or not. If the word doesn't fit on the
current line but fits on the next one, simply put it there. Otherwise
split the word at the end of the line and continue on the next line.